+++ raspberry-io-pi4j/ADC/src/main/python/adcMCP3008_py2_1.py
@@ -44,7 +44,7 @@ while keep_looping:
     try:
         value = chan.value
         voltage = chan.voltage
-        adc = (voltage / 3.3) * 1023
+        adc = int((voltage / 3.3) * 1023)
 #         print('Raw ADC Value: ', value)
 #         print('ADC Voltage: ' + str(voltage) + 'V')
         print("Raw: {}, Voltage: {}, ADC: {}".format(value, str(voltage), adc))   # Look into {adc: %d}...
